Add ScheduleDaysService Modified ScheduleDay Views

+++ Services/CustomERP.Services.Data/SchedulesService.cs
@@ -17,11 +17,15 @@
             this.scheduleRepository = scheduleRepository;
         }
 
-        public void Delete(int id)
+        public async Task Delete(int id, string creator)
         {
             var schedule = this.scheduleRepository.All().FirstOrDefault(x => x.Id == id);
-            this.scheduleRepository.Delete(schedule);
-            this.scheduleRepository.SaveChangesAsync();
+            if (schedule != null)
+            {
+                schedule.DeletedFrom = creator;
+                this.scheduleRepository.Delete(schedule);
+                await this.scheduleRepository.SaveChangesAsync();
+            }
         }
 
         public async Task<int> Update(int id, string name, int numberOfDays, string modifier)
@@ -54,26 +58,48 @@
             return schedule.Id;
         }
 
-        public bool AnyByName<T>(string name)
+        public bool AnyByName<T>(string name, int? id)
         {
-            var schedule = this.scheduleRepository.All().Where(x => x.Name == name).To<T>().FirstOrDefault();
-            return schedule == null;
+            var schedule = this.scheduleRepository
+                .All().Where(x => x.Name == name && x.Id != id)
+                .To<T>().FirstOrDefault();
+            return schedule != null;
         }
 
         public T GetById<T>(int? id)
         {
-            var schedule = this.scheduleRepository.All().Where(s => s.Id == id).To<T>().FirstOrDefault();
+            var schedule = this.scheduleRepository
+                .All()
+                .Where(s => s.Id == id)
+                .To<T>().FirstOrDefault();
             return schedule;
         }
 
         public IEnumerable<T> GetAll<T>()
         {
-            return this.scheduleRepository.All().To<T>().ToList();
+            return this.scheduleRepository
+                .All().To<T>().ToList();
         }
 
         public int GetCount()
         {
             return this.scheduleRepository.All().Count();
+        }
+
+        public int GetNumberOfDaysById(int id)
+        {
+            return this.scheduleRepository
+                .All()
+                .Where(x => x.Id == id)
+                .Select(x => x.NumberOfDays)
+                .FirstOrDefault();
+        }
+
+        public bool AnyById(int id)
+        {
+            var result = this.scheduleRepository.All().FirstOrDefault(x => x.Id == id);
+
+            return result != null;
         }
     }
 }
